feat(analyzer): add files
