+++ analyzer/config/config.go
@@ -5,16 +5,13 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/aquasecurity/fanal/analyzer/config/cloudformation"
 	"golang.org/x/xerrors"
 
 	"github.com/aquasecurity/fanal/analyzer"
 	"github.com/aquasecurity/fanal/analyzer/config/dockerfile"
-	"github.com/aquasecurity/fanal/analyzer/config/kubernetes"
+	"github.com/aquasecurity/fanal/analyzer/config/json"
 	"github.com/aquasecurity/fanal/analyzer/config/terraform"
-	"github.com/aquasecurity/fanal/analyzer/config/toml"
 	"github.com/aquasecurity/fanal/analyzer/config/yaml"
-	"github.com/aquasecurity/fanal/config/scanner"
 	"github.com/aquasecurity/fanal/types"
 )
 
@@ -35,9 +32,9 @@ func (o *ScannerOption) Sort() {
 	sort.Strings(o.DataPaths)
 }
 
-func RegisterConfigAnalyzers(rootPath string, opt ScannerOption) error {
-	var dockerRegexp, k8sRegexp, tomlRegexp, yamlRegexp *regexp.Regexp
-	for _, p := range opt.FilePatterns {
+func RegisterConfigAnalyzers(filePatterns []string) error {
+	var dockerRegexp, jsonRegexp, yamlRegexp *regexp.Regexp
+	for _, p := range filePatterns {
 		// e.g. "dockerfile:my_dockerfile_*"
 		s := strings.SplitN(p, separator, 2)
 		if len(s) != 2 {
@@ -52,10 +49,8 @@ func RegisterConfigAnalyzers(rootPath string, opt ScannerOption) error {
 		switch fileType {
 		case types.Dockerfile:
 			dockerRegexp = r
-		case types.Kubernetes:
-			k8sRegexp = r
-		case types.TOML:
-			tomlRegexp = r
+		case types.JSON:
+			jsonRegexp = r
 		case types.YAML:
 			yamlRegexp = r
 		default:
@@ -63,16 +58,9 @@ func RegisterConfigAnalyzers(rootPath string, opt ScannerOption) error {
 		}
 	}
 
-	s, err := scanner.New(rootPath, opt.Namespaces, opt.PolicyPaths, opt.DataPaths, opt.Trace)
-	if err != nil {
-		return xerrors.Errorf("scanner init error: %w", err)
-	}
-
-	analyzer.RegisterPostAnalyzer(dockerfile.NewPostAnalyzer(s, dockerRegexp))
-	analyzer.RegisterPostAnalyzer(kubernetes.NewPostAnalyzer(s, k8sRegexp))
+	analyzer.RegisterAnalyzer(dockerfile.NewConfigAnalyzer(dockerRegexp))
 	analyzer.RegisterAnalyzer(terraform.NewConfigAnalyzer())
-	analyzer.RegisterAnalyzer(cloudformation.NewConfigAnalyzer())
-	analyzer.RegisterAnalyzer(toml.NewConfigAnalyzer(tomlRegexp))
+	analyzer.RegisterAnalyzer(json.NewConfigAnalyzer(jsonRegexp))
 	analyzer.RegisterAnalyzer(yaml.NewConfigAnalyzer(yamlRegexp))
 
 	return nil
