+++ analyzer/config/terraform/terraform.go
@@ -2,8 +2,12 @@ package terraform
 
 import (
 	"context"
+	"io"
 	"os"
 	"path/filepath"
+
+	"golang.org/x/exp/slices"
+	"golang.org/x/xerrors"
 
 	"github.com/aquasecurity/fanal/analyzer"
 	"github.com/aquasecurity/fanal/types"
@@ -11,7 +15,7 @@ import (
 
 const version = 1
 
-const requiredExt = ".tf"
+var requiredExts = []string{".tf", ".tf.json"}
 
 type ConfigAnalyzer struct {
 }
@@ -22,18 +26,26 @@ func NewConfigAnalyzer() ConfigAnalyzer {
 
 // Analyze returns a name of Terraform file
 func (a ConfigAnalyzer) Analyze(_ context.Context, input analyzer.AnalysisInput) (*analyzer.AnalysisResult, error) {
+	b, err := io.ReadAll(input.Content)
+	if err != nil {
+		return nil, xerrors.Errorf("read error (%s): %w", input.FilePath, err)
+	}
 	return &analyzer.AnalysisResult{
-		Misconfigurations: []types.Config{
-			{
-				Type:     types.Terraform,
-				FilePath: filepath.Join(input.Dir, input.FilePath), // tfsec requires a path from working dir
+		Files: map[types.HandlerType][]types.File{
+			// It will be passed to misconf post handler
+			types.MisconfPostHandler: {
+				{
+					Type:    types.Terraform,
+					Path:    filepath.Join(input.Dir, input.FilePath), // tfsec requires a path from working dir
+					Content: b,
+				},
 			},
 		},
 	}, nil
 }
 
 func (a ConfigAnalyzer) Required(filePath string, _ os.FileInfo) bool {
-	return filepath.Ext(filePath) == requiredExt
+	return slices.Contains(requiredExts, filepath.Ext(filePath))
 }
 
 func (ConfigAnalyzer) Type() analyzer.Type {
