+++ handler/all/import.go
@@ -1,0 +1,6 @@
+package all
+
+import (
+	_ "github.com/aquasecurity/fanal/handler/misconf"
+	_ "github.com/aquasecurity/fanal/handler/sysfile"
+)
